ExactDiagonalization module
