+++ src/DictVectors/DictVectors.jl
@@ -20,8 +20,7 @@ import MPI
 using ..Interfaces: Interfaces, AbstractDVec, AdjointUnknown,
     CompressionStrategy, IsDiagonal, LOStructure,
     apply_column!, apply_operator!, compress!,
-    diagonal_element, offdiagonals, step_stats
-using ..Hamiltonians: Hamiltonians, AbstractHamiltonian
+    diagonal_element, offdiagonals, step_stats, AbstractHamiltonian
 using ..StochasticStyles: StochasticStyles, IsDeterministic
 
 import ..Interfaces: deposit!, storage, StochasticStyle, default_style, freeze, localpart,
