+++ test/Hamiltonians.jl
@@ -1289,133 +1289,6 @@ end
 
 end
 
-@testset "BasisSetRep" begin
-    @testset "basics" begin
-        m = 100
-        n = 100
-        addr = BoseFS(Tuple(i == 1 ? n : 0 for i in 1:m))
-        ham = HubbardReal1D(addr)
-        @test_throws ArgumentError BasisSetRep(ham) # dimension too large
-        m = 2
-        n = 10
-        addr = near_uniform(BoseFS{n,m})
-        ham = HubbardReal1D(addr)
-        bsr = BasisSetRep(ham; nnzs = dimension(ham))
-        @test length(bsr.basis) == dimension(bsr) ≤  dimension(ham)
-        @test_throws ArgumentError BasisSetRep(ham, BoseFS((1,2,3))) # wrong address type
-        @test Matrix(bsr) == Matrix(bsr.sm) == Matrix(ham)
-        @test sparse(bsr) == bsr.sm == sparse(ham)
-        addr2 = bsr.basis[2]
-        @test starting_address(BasisSetRep(ham, addr2)) ==  addr2
-        @test isreal(ham) == (eltype(ham) <: Real)
-        @test isdiag(ham) == (LOStructure(ham) ≡ IsDiagonal())
-        @test ishermitian(ham) == (LOStructure(ham) ≡ IsHermitian())
-        @test issymmetric(ham) == (ishermitian(ham) && isreal(ham))
-    end
-
-    @testset "filtering" begin
-        ham = HubbardReal1D(near_uniform(BoseFS{10,2}))
-        bsr_orig = BasisSetRep(ham; sort=true)
-        mat_orig = Matrix(bsr_orig)
-        mat_cut_index = diag(mat_orig) .< 30
-        mat_cut_manual = mat_orig[mat_cut_index, mat_cut_index]
-        bsr = BasisSetRep(ham; cutoff=30, sort=true)
-        mat_cut = Matrix(bsr)
-        @test mat_cut == mat_cut_manual
-        # pass a basis and generate truncated BasisSetRep
-        bsrt = BasisSetRep(ham, bsr.basis; filter= Returns(false), sort=true)
-        @test bsrt.basis == bsr.basis
-        @test bsr.sm == bsrt.sm
-        # pass addresses and generate reachable basis
-        @test BasisSetRep(ham, bsr.basis, sort=true).basis == bsr_orig.basis
-
-        filterfun(fs) = maximum(onr(fs)) < 8
-        mat_cut_index = filterfun.(BasisSetRep(ham; sort=true).basis)
-        mat_cut_manual = mat_orig[mat_cut_index, mat_cut_index]
-        mat_cut = Matrix(ham; filter=filterfun, sort=true)
-        @test mat_cut == mat_cut_manual
-    end
-
-    @testset "getindex" begin
-        ham = HubbardReal1D(near_uniform(BoseFS{10,2}))
-        bsr = BasisSetRep(ham; sort=true)
-        b = bsr.basis
-        @test [ham[i, j] for i in b, j in b] == Matrix(bsr)
-    end
-
-    @testset "momentum blocking" begin
-        add1 = BoseFS((2,0,0,0))
-        add2 = BoseFS((0,1,0,1))
-        ham = HubbardMom1D(add1)
-
-        @test Matrix(ham, add1; sort=true) == Matrix(ham, add2; sort=true)
-        @test Matrix(ham, add1) ≠ Matrix(ham, add2)
-
-        add1 = BoseFS((2,0,0,0,0))
-        add2 = BoseFS((0,1,0,0,1))
-        ham = HubbardMom1D(add1)
-
-        @test Matrix(ham, add1; sort=true) == Matrix(ham, add2; sort=true)
-        @test Matrix(ham, add1) ≠ Matrix(ham, add2)
-    end
-
-    using Rimu.Hamiltonians: fix_approx_hermitian!, isapprox_enforce_hermitian!
-    using Rimu.Hamiltonians: build_sparse_matrix_from_LO
-    using Random
-    @testset "fix_approx_hermitian!" begin
-        # generic `Matrix`
-        Random.seed!(17)
-        mat = rand(5,5)
-        @test !ishermitian(mat)
-        @test_throws ArgumentError fix_approx_hermitian!(mat; test_approx_symmetry=true)
-        @test !ishermitian(mat) # still not hermitian
-        fix_approx_hermitian!(mat; test_approx_symmetry=false)
-        @test ishermitian(mat) # now it is hermitian
-
-        # sparse matrix
-        Random.seed!(17)
-        mat = sparse(rand(5,5))
-        @test !ishermitian(mat)
-        @test_throws ArgumentError fix_approx_hermitian!(mat; test_approx_symmetry=true)
-        @test !ishermitian(mat) # still not hermitian
-
-        # subtle symmetry violation due to `ParitySymmetry` wrapper
-        ham = HubbardMom1D(BoseFS((1, 0, 1, 2, 0)))
-        even = ParitySymmetry(ham; odd=false)
-        odd = ParitySymmetry(ham; even=false)
-
-        even_sm, _ = build_sparse_matrix_from_LO(even)
-        even_m = Matrix(even) # symmetrised version via BasisSetRep
-
-        @test !issymmetric(even_sm) # not symmetric due to floating point errors
-        @test issymmetric(even_m) # because it was passed through `fix_approx_hermitian!`
-        @test even_sm ≈ even_m # still approximately the same!
-    end
-
-    @testset "basis-only" begin
-        m = 5
-        n = 5
-        add = near_uniform(BoseFS{n,m})
-        ham = HubbardReal1D(add)
-        @test_throws ArgumentError build_basis(ham, BoseFS((1,2,3))) # wrong address type
-        # same basis as BSR
-        bsr = BasisSetRep(ham)
-        basis = build_basis(ham)
-        @test basis == bsr.basis
-        @test basis == build_basis(ham, basis) # passing multiple addresses
-        # sorting
-        basis = build_basis(ham, add; sort = true)
-        @test basis == sort!(bsr.basis)
-        # filtering
-        @test_throws ArgumentError build_basis(ham, add; sizelim = 100)
-        @test length(build_basis(ham, add; cutoff = -1)) == 1 # no new addresses added
-        cutoff = n * (n-1) / 4  # half maximum energy
-        bsr = BasisSetRep(ham, add; cutoff)
-        basis = build_basis(ham, add; cutoff)
-        @test basis == bsr.basis
-    end
-end
-
 @testset "Stoquastic" begin
     ham = HubbardMom1D(BoseFS((0,5,0))) # a Hamiltonian that has a sign problem
     sham = Stoquastic(ham) # sign problem removed, but smaller ground state eigenvalue
